refactor dex file layout
