+++ contracts/dex/src/handlers/execute.rs
@@ -1,7 +1,6 @@
 use crate::error::DexError;
-use crate::providers::exchange_resolver;
+use crate::exchanges::exchange_resolver;
 use crate::msg::{DexAction, DexExecuteMsg, DexName, IBC_DEX_ID};
-use crate::LocalDex;
 use crate::{
     contract::{DexAdapter, DexResult},
     state::SWAP_FEE,
@@ -73,7 +72,12 @@ fn handle_local_request(
     exchange: String,
 ) -> DexResult {
     let exchange = exchange_resolver::resolve_exchange(&exchange)?;
-    let (msgs, _) = adapter.resolve_dex_action(deps.as_ref(), action, exchange)?;
+    let (msgs, _) = crate::traits::adapter::DexAdapter::resolve_dex_action(
+        &adapter,
+        deps.as_ref(),
+        action,
+        exchange,
+    )?;
     let proxy_msg = adapter.executor(deps.as_ref()).execute(msgs)?;
     Ok(Response::new().add_message(proxy_msg))
 }
