+++ contracts/dex/src/traits/identity.rs
@@ -1,0 +1,4 @@
+pub trait Identify {
+    fn over_ibc(&self) -> bool;
+    fn name(&self) -> &'static str;
+}
